fix: fix undefined is not an object (evaluating 'lastQuizTime.time')
In this commit I've tried to fix the error reported by the Udacity reviewer. I wasn't able to reproduce the issue, so it's kind of a blind try.
In the meanwhile i tested the app more in depth, added iOS support and tested it with iOS too, expanded the readme, and added some comments to the notifications.js file that might have lacked some explanations